refactor(dev): complete Dev entrypoint package split (phases 2–5)
Extract remaining god-module slices into images/common/devcake_dev/:
- harness.render / tokens / argv (live relay, dumps, CLI argv)
- workspace.* (clone, skills/MCP, forensics, activity, transcript)
- adapters.bus (Redis send, artifacts, request-reply, heartbeat)

/dev_entrypoint.py is composition + re-exports (~600 lines). Tests that
patched façade send/time/MAX_ARTIFACT_BYTES now target the owning modules.
docs/07 documents the package map.
