+++ docs/07-dev-runtime.md
@@ -167,7 +167,14 @@ Git pushes and PR interactions (EXECUTE, REVIEW approval checkout) happen inside
 
 **`devcake-relay` is not a shipped CLI.** Devs speak the Redis protocol of `09-messaging.md` **directly from the shared entrypoint** (`runspec.get`, `activity.get`, heartbeats, artifacts). There is no separate relay binary in the image.
 
-**Entrypoint layout.** The image `ENTRYPOINT` is still `/dev_entrypoint.py` (façade). Pure Dev-side logic lives under `images/common/devcake_dev/` (copied to `/devcake_dev` in harness images) — a small Zone-B package separate from `app/devcake`. Phase 1 of the split: `devcake_dev.domain.fault` holds ADR-0018 harness fault classification; further workspace/relay/bus slices may follow.
+**Entrypoint layout.** The image `ENTRYPOINT` is still `/dev_entrypoint.py` (composition façade). Dev-side code lives under `images/common/devcake_dev/` (copied to `/devcake_dev`) — a Zone-B package separate from `app/devcake`:
+
+| Package path | Role |
+|---|---|
+| `domain.fault` | ADR-0018 harness fault classification (pure) |
+| `harness.render` / `tokens` / `argv` | Live relay, stream dumps, CLI argv |
+| `workspace.*` | Clone, skills/MCP, forensics, activity, transcript |
+| `adapters.bus` | Redis Streams send / request-reply / artifacts |
 
 There is **no write access** to the PMO mid-run (INV-4). Writes travel as end-of-run artifacts that the app applies. Mid-run re-fetch of the activity payload uses the same `activity.get` request/reply channel the entrypoint already speaks.
