+++ images/common/devcake_dev/adapters/__init__.py
@@ -1,0 +1,1 @@
+"""I/O adapters for the Dev process."""
