+++ images/common/devcake_dev/workspace/__init__.py
@@ -1,0 +1,1 @@
+"""Workspace prep and activity materialization."""
